@@ -3,14 +3,14 @@ plugins {
 }
 
 android {
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.aniketjain.weatherapp"
         minSdk 19
-        targetSdk 32
-        versionCode 4
-        versionName "1.0.4"
+        targetSdk 33
+        versionCode 5
+        versionName "1.0.5"
         multiDexEnabled true
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -18,13 +18,11 @@ android {
 
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-    buildFeatures {
-        viewBinding true
-    }
+    buildFeatures { viewBinding true }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -32,18 +30,17 @@ android {
 }
 
 dependencies {
-
-    implementation 'androidx.appcompat:appcompat:1.5.0'
-    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.9.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     //for getting json data
     implementation 'com.android.volley:volley:1.2.1'
     //location - lat, lon
-    implementation 'com.google.android.gms:play-services-location:20.0.0'
+    implementation 'com.google.android.gms:play-services-location:21.0.1'
     //custom toast
     implementation 'com.github.dev-aniketj:roasted-toast:1.0.2'
     //custom progress
@@ -56,6 +53,6 @@ dependencies {
     implementation 'com.intuit.sdp:sdp-android:1.1.0'
     //Lotti files
     implementation 'com.airbnb.android:lottie:5.2.0'
-
+    // Multi Dex Enable
     implementation 'androidx.multidex:multidex:2.0.1'
 }
