@@ -18,7 +18,6 @@ import android.view.View;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputMethodManager;
 
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
@@ -93,8 +92,8 @@ public class HomeActivity extends AppCompatActivity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == REQUEST_CODE_EXTRA_INPUT){
-            if(resultCode == RESULT_OK && data!=null){
+        if (requestCode == REQUEST_CODE_EXTRA_INPUT) {
+            if (resultCode == RESULT_OK && data != null) {
                 ArrayList<String> arrayList = data.getStringArrayListExtra(RecognizerIntent.EXTRA_RESULTS);
                 binding.layout.cityEt.setText(Objects.requireNonNull(arrayList).get(0).toUpperCase());
                 searchCity(binding.layout.cityEt.getText().toString());
@@ -147,19 +146,16 @@ public class HomeActivity extends AppCompatActivity {
             binding.mainRefreshLayout.setRefreshing(false);  //for the next time
         });
         //Mic Search
-        binding.layout.micSearchId.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH);
-                intent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL,RecognizerIntent.LANGUAGE_MODEL_FREE_FORM);
-                intent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, Locale.getDefault());
-                intent.putExtra(RecognizerIntent.EXTRA_PROMPT,REQUEST_CODE_EXTRA_INPUT);
-                try {
-                    //it was deprecated but still work
-                    startActivityForResult(intent,REQUEST_CODE_EXTRA_INPUT);
-                }catch (Exception e){
-                    Log.d("Error Voice", "Mic Error:  "+e);
-                }
+        binding.layout.micSearchId.setOnClickListener(view -> {
+            Intent intent = new Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH);
+            intent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM);
+            intent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, Locale.getDefault());
+            intent.putExtra(RecognizerIntent.EXTRA_PROMPT, REQUEST_CODE_EXTRA_INPUT);
+            try {
+                //it was deprecated but still work
+                startActivityForResult(intent, REQUEST_CODE_EXTRA_INPUT);
+            } catch (Exception e) {
+                Log.d("Error Voice", "Mic Error:  " + e);
             }
         });
     }
